ADD:** special find functions
These functions are not used anymore, but might come in handy. 
Goal was to use these with the find modules, if modules was not found then they would be fetched. When fetch they could be installed and if installed we would not do another time consuming check... 
This was mainly a fix for windows since it was slow at using fetchContent...