FIX: resource manager
Fixed some problems relating to the resource manager. Decided path in the config files were always used and should not be the case when programming in the Game repo.

Also changed it so wrong paths gave a warning message and that it after used  the default mesh/texture. Before it threw runtime errors.